+++ app/src/main/res/layout/activity_training_statistics.xml
@@ -162,7 +162,7 @@
             <FrameLayout
                 android:id="@+id/time_on_each_slide_chart_box_activity_training_statistics"
                 android:layout_width="match_parent"
-                android:layout_height="250dp"
+                android:layout_height="wrap_content"
                 android:layout_marginVertical="10dp" />
 
             <View
